Move exceptions to AudioChord.Exceptions
