+++ src/AudioChord/MusicService.cs
@@ -8,6 +8,7 @@ using MongoDB.Driver;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using AudioChord.Exceptions;
 
 namespace AudioChord
 {
